requirements.txt
streamlit
openai
sentence-transformers
python-dotenv
torch
langchain
faiss-cpu
rank-bm25